% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addLogo}
\alias{addLogo}
\title{Add default logo to the report}
\usage{
addLogo(htmlObj)
}
\arguments{
\item{htmlObj}{S3 object created by initHTML()}
}
\value{
updated htmlObj
}
\description{
Add default logo (Gilead Bioinformatics) to the report
}

