% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addReport}
\alias{addReport}
\title{Add a report content}
\usage{
addReport(report.path, htmlObj)
}
\arguments{
\item{report.path}{path to local report file (must be written in markdown syntax)}

\item{htmlObj}{S3 object created by initHTML()}
}
\value{
updated htmlObj
}
\description{
Add report content (such as Introduction, Discussion) to htmlObj
}
\details{
Report file must be written in markdown syntax so that the program can parse
	and write it into .html format. For quick start, please refer to the template file
 provided by this package. For more details, please go to \url{http://daringfireball.net/projects/markdown/syntax}
}

