% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertToJson}
\alias{convertToJson}
\alias{convertToJson.default}
\title{convert input to a javascript object(json)}
\usage{
convertToJson(x, ...)

\method{convertToJson}{default}(x, outDir = ".", outPrefix = NULL,
  fieldnames = NULL)
}
\arguments{
\item{x}{input}

\item{...}{other arguments}

\item{outDir}{path Path to save the .json file}

\item{outPrefix}{character prefix of json file}

\item{fieldnames}{vector Column names of the data table}
}
\description{
convert input into a javascript object and write to a file

convert to json
}

