\name{DEgene.df}
\alias{DEgene.df}
\docType{data}
\title{
Differentially expressed gene list
}
\description{
Differentially expressed gene list created by EdgeR
}
\usage{data("DEgene.df")}
\format{
  A data frame with 14145 observations on the following 4 variables.
  \describe{
    \item{\samp{Gene name}}{a character vector}
    \item{\samp{log2 fold change}}{a character vector}
    \item{\samp{p-value}}{a character vector}
    \item{\code{fdr}}{a character vector}
  }
}
