\name{gwas.df}
\alias{gwas.df}
\docType{data}
\title{
gwas sample data
}
\description{
gwas sample data
}
\usage{data("gwas.df")}
\format{
  A data frame with 10000 observations on the following 5 variables.
  \describe{
    \item{\code{chr}}{a numeric vector}
    \item{\code{SNPName}}{a character vector}
    \item{\code{pos}}{a numeric vector}
    \item{\code{PValue}}{a numeric vector}
    \item{\samp{-logP}}{a numeric vector}
  }
}
