% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initHTML}
\alias{initHTML}
\title{Initiate an htmlObj}
\usage{
initHTML(title = "My Report", outDir = ".", outPrefix = "htmlReport",
  author = "", date = "")
}
\arguments{
\item{title}{character Report title}

\item{outDir}{path Path of output files}

\item{outPrefix}{character The string used for creating output folder}

\item{author}{character Report author}

\item{date}{character Report date}
}
\value{
htmlObj
}
\description{
create new htmlObj, and set up folders containing required css and js libraries
}

